kgo-verifier: refactor into worker classes
This is a precursor to generic status reads over HTTP.
